[dv] Alter cov_merge.tcl patch so icache coverage collection works
cov_merge.tcl has been patched so the Ibex regression can pass through
coverage directories via a file. However this breaks dvsim based flows,
in particular the icache coverage merge.

This patch checks to see if the file is provided, if not it uses the
original merge method that's compatible with dvsim.