Formal verification flow for CHERIoT Ibex
This commit adds the content of the dv/formal folder. Please look at the
README in there for more details on how to use this. Here's a bullet
list of what is included in this commit:
- instrs.json that includes RISC-V and CHERIoT instructions.
- spec_api that is capability aware
- Pipeline follower.
- Comparison utilities between Sail and Ibex.
- Data type invariant for CHERIoT capabilities.
- Alternative LSUs so that we don't need an uperbound on memory response
  time.
- Temporal safety, enabling both trvk and tbre (although revocation
  engine is never started).
- Nix development environment.
- Interrupt verification including non-maskable and fast ones.

Some additional more detailed notes are captured below:

We maintain an invariant throughout the entirety of memory
operations equivalent to the statement "if the memory operation
were to finish now, it would do so correctly". This is done with
several alternative LSUs.
- The first is active before the first rvalid from memory, while in
  EX.
- The second is active after the first rvalid from memory, also while
  in EX.
- The third is active throughout WB.
It's possible that these are not really necessary, and instead a
stronger invariant about rdata_q etc could be used, but this way
is at least intuitive.

MSHWM and MSHWMB are disabled (i.e. not checked at all).

The majority prove with Hp and should be run individually, apart from
Step0 and various specific examples. Some prove at a higher proof
effort than I expect.

Some properties are somewhat problematic: Step23::top.SpecPastNoWbexcPC
in particular, but this seems to be a tool bug. Querying an SST trace
of length 5 fails to find anything, but Hp cannot prove it. Further,
prove -sst 5 find a trace, but attempting to load it gives a warning
saying an assumption has been violated, meaning the trace was not a
trace at all. It is nonetheless eventually proved by D, but it
ultimately should be proved much easier.

Better proof script, which actually does some stuff in parrallel,
especially the bulky middle stuff which is now many individual Hp
instances. This should make automated proofs a fair bit faster.

PC properties now prove way faster, especially MRET PC.
Initial_IdleActive proves fast regardless of the bug now.
CSR_CSR_NoErr proves fast.

Co-authored-by: Harry Callahan <hcallahan@lowrisc.org>
Co-authored-by: Marno van der Maas <mvdmaas+git@lowrisc.org>